refactor(core): add recovery/initiate, getBip32Keys
The `initiateRecovery` method has an unfortunate signature which
cannot be improved easily.

In addition, there is a lot of duplication between subclassed
overrides that should be cleaned up.

Add helper method `getBip32Keys(InitiateRecoveryOptions)` to reduce
duplication.

Issue: BG-34381
